[FE] refactor : ModalProvider 컴포넌트 Service 영역으로 이전
