+++ frontend/src/App.tsx
@@ -6,18 +6,22 @@ import { QueryClientProvider } from '@tanstack/react-query';
 import PageRoutes from 'PageRoutes';
 import { RecoilRoot } from 'recoil';
 
-import { ErrorBoundary, ModalProvider, SnackbarProvider } from 'common/components';
+import { ErrorBoundary, SnackbarProvider } from 'common/components';
+
+import ModalProvider from 'service/@shared/components/ModalProvider';
 
 import queryClient from 'api/config/queryClient';
 import ErrorPage from 'service/@shared/pages/ErrorPage';
-import * as modalContentList from 'service/@shared/pages/modals';
 import 'styles/@app.scss';
 
 function ContextWrapper({ children }: { children: ReactNode }) {
   return (
     <RecoilRoot>
       <BrowserRouter>
-        <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+        <QueryClientProvider client={queryClient}>
+          <ModalProvider />
+          {children}
+        </QueryClientProvider>
       </BrowserRouter>
     </RecoilRoot>
   );
@@ -28,8 +32,6 @@ function App() {
     <ContextWrapper>
       <SnackbarProvider />
       <Suspense>
-        <ModalProvider contentList={modalContentList} />
-
         <ErrorBoundary fallback={ErrorPage}>
           <PageRoutes />
         </ErrorBoundary>
