+++ frontend/src/common/components/index.ts
@@ -13,7 +13,6 @@ export { default as Snackbar } from './Snackbar';
 export { default as SnackbarProvider } from './SnackbarProvider';
 export { default as PopupBox } from './PopupBox';
 export { default as Modal } from './Modal';
-export { default as ModalProvider } from './ModalProvider';
 export { default as FlexContainer } from './FlexContainer';
 export { default as PaginationBar } from './PaginationBar';
 export { default as ToolTip } from './ToolTip';
